+++ derive-input/src/internal/column.rs
@@ -1,10 +1,13 @@
 use crate::api::{
     Column,
     db::{SpacetimeDBColumn, SpacetimeDBTable},
-    dsl::{method::SpacetimeDSLColumnMethods, table::SpacetimeDSLTable},
+    dsl::{
+        column::SpacetimeDSLColumn, method::SpacetimeDSLColumnMethods, table::SpacetimeDSLTable,
+    },
     rust::{RustField, RustStruct},
 };
 use spacetime_bindings_macro_input::table::ColumnArgs;
+use syn::DeriveInput;
 
 mod rust;
 
@@ -13,17 +16,14 @@ mod db;
 mod dsl;
 
 pub(in crate::internal) fn try_parse(
-    item: &syn::DeriveInput,
+    item: &DeriveInput,
     column_args: &ColumnArgs,
     rust_struct: &RustStruct,
     mut spacetimedb_table: SpacetimeDBTable,
     mut spacetimedsl_table: SpacetimeDSLTable,
 ) -> syn::Result<(SpacetimeDBTable, SpacetimeDSLTable, Vec<Column>)> {
-    let sequenced_columns = &columns_to_string(&column_args.sequenced_columns);
-    let primary_key_column = &column_args
-        .primary_key_column
-        .as_ref()
-        .map(|c| column_to_string(c));
+    let auto_inc_column_names = get_auto_inc_column_names(column_args);
+    let primary_key_column_name = get_primary_key_column_name(column_args);
 
     let mut columns = vec![];
 
@@ -31,20 +31,15 @@ pub(in crate::internal) fn try_parse(
         let rust_field = RustField::map(field);
 
         let res = SpacetimeDBColumn::map(
+            &rust_field,
             spacetimedb_table,
-            primary_key_column,
-            sequenced_columns,
-            field,
+            &auto_inc_column_names,
+            &primary_key_column_name,
         );
         spacetimedb_table = res.0;
         let spacetimedb_column = res.1;
 
-        let res = crate::api::dsl::column::SpacetimeDSLColumn::try_parse(
-            item,
-            field,
-            &rust_field,
-            spacetimedsl_table,
-        )?;
+        let res = SpacetimeDSLColumn::try_parse(item, field, &rust_struct, &rust_field, spacetimedsl_table)?;
         spacetimedsl_table = res.0;
         let spacetimedsl_column = res.1;
 
@@ -68,14 +63,17 @@ pub(in crate::internal) fn try_parse(
     Ok((spacetimedb_table, spacetimedsl_table, columns))
 }
 
-pub(in crate::internal) fn columns_to_string(
-    columns: &Vec<spacetime_bindings_macro_input::table::Column<'_>>,
-) -> Vec<String> {
-    columns.iter().map(|c| column_to_string(c)).collect()
+fn get_auto_inc_column_names(column_args: &ColumnArgs<'_>) -> Vec<Box<str>> {
+    column_args
+        .sequenced_columns
+        .iter()
+        .map(|c| c.ident.to_string().into())
+        .collect()
 }
 
-pub(in crate::internal) fn column_to_string(
-    column: &spacetime_bindings_macro_input::table::Column<'_>,
-) -> String {
-    column.ident.to_string()
+fn get_primary_key_column_name(column_args: &ColumnArgs<'_>) -> Option<Box<str>> {
+    column_args
+        .primary_key_column
+        .as_ref()
+        .map(|c| c.ident.to_string().into())
 }
