+++ derive-input/src/internal/column/rust.rs
@@ -1,0 +1,17 @@
+use crate::api::rust::{RustField, RustVisibility};
+use quote::ToTokens;
+use spacetime_bindings_macro_input::sats::SatsField;
+
+impl RustField {
+    pub(in crate::internal) fn map(field: &SatsField<'_>) -> RustField {
+        let visibility = RustVisibility::map(field.vis);
+        let name = field.ident.as_ref().unwrap().to_string().into();
+        let type_name_or_path = field.ty.to_token_stream().to_string().into();
+
+        RustField {
+            visibility,
+            name,
+            type_name_or_path,
+        }
+    }
+}
