+++ derive-input/src/internal/dsl.rs
@@ -1,6 +1,4 @@
-pub mod table;
-
-pub mod column;
+use spacetime_bindings_macro_input::{sym::Symbol, symbol};
 
 pub mod wrapper;
 
@@ -13,3 +11,13 @@ pub mod setter;
 pub mod method;
 
 pub mod quote;
+
+symbol!(foreign_key);
+symbol!(table);
+symbol!(column);
+symbol!(on_delete);
+symbol!(wrap);
+symbol!(wrapped);
+symbol!(plural_name);
+symbol!(unique_index);
+symbol!(path);
