--- derive-input/src/internal/dsl/column.rs
@@ -1,42 +1,0 @@
-use super::{foreign_key, getter, setter};
-use crate::api::{dsl::{column::SpacetimeDSLColumn, table::SpacetimeDSLTable, wrapper::WrapperType}, rust::RustField};
-use quote::ToTokens;
-use spacetime_bindings_macro_input::sats::SatsField;
-
-impl SpacetimeDSLColumn {
-    pub(in crate::internal) fn try_parse(
-        item: &syn::DeriveInput,
-        field: &SatsField<'_>,
-        rust_field: &RustField,
-        mut spacetimedsl_table: SpacetimeDSLTable,
-    ) -> syn::Result<(SpacetimeDSLTable, SpacetimeDSLColumn)> {
-        let is_option = field
-            .ty
-            .to_token_stream()
-            .to_string()
-            .starts_with("Option<");
-
-        let wrapper_type = WrapperType::try_parse(item, field)?;
-
-        let foreign_key = foreign_key::try_parse(field)?;
-
-        let getter = getter::get_getter(rust_field, is_option, &wrapper_type);
-
-        let setter = setter::get_setter(rust_field, is_option, &wrapper_type);
-
-        if setter.is_some() {
-            spacetimedsl_table.is_mutable = true;
-        }
-
-        Ok((
-            spacetimedsl_table,
-            SpacetimeDSLColumn {
-                is_option,
-                wrapper_type,
-                foreign_key,
-                getter,
-                setter,
-            },
-        ))
-    }
-}
