+++ derive-input/src/internal/dsl/method/create.rs
@@ -1,3 +1,4 @@
+use crate::internal::utils::wrapper_type_into_option;
 use crate::{
     api::{
         Column,
@@ -5,12 +6,9 @@ use crate::{
         dsl::{method::SpacetimeDSLMethod, wrapper::WrapperType},
         rust::RustStruct,
     },
-    internal::{
-        dsl::quote::{
-            get_column_value, get_column_value_from_wrapper, get_method_arg_column_type,
-            get_method_arg_into_wrapper_type, get_method_arg_into_wrapper_type_option,
-        },
-        wrapper_type_into_option,
+    internal::dsl::quote::{
+        get_column_value, get_column_value_from_wrapper, get_method_arg_column_type,
+        get_method_arg_into_wrapper_type, get_method_arg_into_wrapper_type_option,
     },
 };
 use proc_macro2::TokenStream;
