+++ derive-input/src/internal/table.rs
@@ -1,8 +1,10 @@
 use crate::api::{
     Table,
+    db::SpacetimeDBTable,
     dsl::{method::SpacetimeDSLTableMethods, table::SpacetimeDSLTable},
 };
 use spacetime_bindings_macro_input::table::{ColumnArgs, TableArgs};
+use syn::DeriveInput;
 
 mod rust;
 
@@ -11,13 +13,13 @@ mod db;
 mod dsl;
 
 pub(in crate::internal) fn try_parse(
-    item: &syn::DeriveInput,
+    item: &DeriveInput,
     table_args: &TableArgs,
     column_args: &ColumnArgs<'_>,
 ) -> syn::Result<Table> {
     let rust_struct = rust::map(&item);
 
-    let spacetimedb_table = db::map(&table_args);
+    let spacetimedb_table = SpacetimeDBTable::map(&table_args);
 
     let (spacetimedb_table, spacetimedsl_table) =
         SpacetimeDSLTable::try_parse(&item, spacetimedb_table)?;
