+++ derive-input/src/internal/table/rust.rs
@@ -1,6 +1,7 @@
 use crate::api::rust::{RustStruct, RustVisibility};
+use syn::DeriveInput;
 
-pub(in crate::internal) fn map(input: &syn::DeriveInput) -> RustStruct {
+pub(in crate::internal) fn map(input: &DeriveInput) -> RustStruct {
     let visibility = RustVisibility::map(&input.vis);
     let name = input.ident.to_string().into();
